banana/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-genai
google-generativeai==0.3.2
google-cloud-aiplatform==1.38.1
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
Pillow==10.1.0
python-alipay-sdk==3.4.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0

